backend/requirements.txt
# ─────────────────────────────────────────────
# Django & REST
# ─────────────────────────────────────────────
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1

# ─────────────────────────────────────────────
# Database & Utilities
# ─────────────────────────────────────────────
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-extensions==3.2.3

# ─────────────────────────────────────────────
# Caching & Task Queue
# ─────────────────────────────────────────────
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# ─────────────────────────────────────────────
# AI / Voice
# ─────────────────────────────────────────────
groq==0.4.2